+++ changelog.markdown
@@ -16,6 +16,7 @@
 + Fixed the functions start() and click() to set the value on target when 'target' option is enabled;
 + Fixed the functions start() and click() not to be executed when 'readOnly' is enabled;
 + Fixed the 'target' option to work with 'half' and 'precision' option;
++ Fixed the index number given to Raty binded by class without id; (fixed by "Remember me, please?")
 + Refactored the code to make it cleanner and faster;
 + Changed the class of the cancel button to 'raty-cancel' to avoid CSS conflicts;
 + Now the scope of click callback is the raty element as DOM, not jQuery selection, to follow the usual;